Moved tx off code from pod interface to dhcp.
